refactor(chat): extract SupportInbox component for reuse
Move all inbox logic from AdminSupportPage into a standalone SupportInbox
component that reads partnerPrepId from the Zustand store to scope Firestore
queries; admin page is now a thin wrapper.
